+++ apps/astarte_pairing/lib/astarte_pairing_web/metrics.ex
@@ -18,6 +18,7 @@
 
 defmodule Astarte.PairingWeb.Metrics do
   def setup do
+    Astarte.PairingWeb.Metrics.PairingInstrumenter.setup()
     Astarte.PairingWeb.Metrics.PrometheusExporter.setup()
   end
 end